Expanded cpuexec cycles <-> attotime functions to handle UINT64s.

+++ src/emu/cpuexec.h
@@ -154,10 +154,10 @@ double cpu_get_clockscale(const device_config *device);
 void cpu_set_clockscale(const device_config *device, double clockscale);
 
 /* converts a number of clock ticks to an attotime */
-attotime cpu_clocks_to_attotime(const device_config *device, UINT32 clocks);
+attotime cpu_clocks_to_attotime(const device_config *device, UINT64 clocks);
 
 /* converts a duration as attotime to CPU clock ticks */
-UINT32 cpu_attotime_to_clocks(const device_config *device, attotime duration);
+UINT64 cpu_attotime_to_clocks(const device_config *device, attotime duration);
